Modify Travis CI script to deploy test version workbench extension to marketplace
