+++ scripts/copyVendor.js
