+++ scripts/modifyConfigIfRCVersion.js
@@ -1,0 +1,22 @@
+const fs = require('fs');
+
+if (process.env.TRAVIS_TAG) {
+  const ISTESTVERSION = /^v?[0-9]+\.[0-9]+\.[0-9]+-[rR][cC]$/.test(process.env.TRAVIS_TAG || '');
+  const packageJson = JSON.parse(fs.readFileSync('package.json'));
+  if (ISTESTVERSION) {
+    // Modify
+    const testName = "test-iot-workbench";
+    const testDisplayName = "Test IoT Workbench";
+    const testPublisher = "IoTDevExBuild";
+    packageJson.name = testName;
+    packageJson.displayName = testDisplayName;
+    packageJson.publisher = testPublisher;
+    packageJson.version = packageJson.version.slice(0, -3);
+
+    // Delete
+    delete packageJson.icon;
+    delete packageJson.aiKey;
+
+    fs.writeFileSync('package.json', JSON.stringify(packageJson, null, 2) + '\n');
+  }
+}
